+++ src/Engines/AlgoliaEngine.php
@@ -7,6 +7,7 @@ use Exception;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\LazyCollection;
 use Laravel\Scout\Builder;
+use Laravel\Scout\Jobs\RemoveableScoutCollection;
 
 class AlgoliaEngine extends Engine
 {
@@ -63,9 +64,9 @@ class AlgoliaEngine extends Engine
             }
 
             return array_merge(
-                ['objectID' => $model->getScoutKey()],
                 $searchableData,
-                $model->scoutMetadata()
+                $model->scoutMetadata(),
+                ['objectID' => $model->getScoutKey()],
             );
         })->filter()->values()->all();
 
@@ -82,13 +83,17 @@ class AlgoliaEngine extends Engine
      */
     public function delete($models)
     {
+        if ($models->isEmpty()) {
+            return;
+        }
+
         $index = $this->algolia->initIndex($models->first()->searchableAs());
 
-        $index->deleteObjects(
-            $models->map(function ($model) {
-                return $model->getScoutKey();
-            })->values()->all()
-        );
+        $keys = $models instanceof RemoveableScoutCollection
+            ? $models->pluck($models->first()->getUnqualifiedScoutKeyName())
+            : $models->map->getScoutKey();
+
+        $index->deleteObjects($keys->all());
     }
 
     /**